refactor: tests/integration/100.command.dev.test.cjs
- convert file to ESM
- replace got with fetch in most of the test cases
- define environment prefix so that vitest can set environment variables
- move and rename file

Related to: #5698
